Set MO_AGC_GAIN_ADJ2 to default values on PCI resume callback
For some reason when the computer is awoken from sleep, the card's
registers aren't in their last state or their default state. This
measn the MO_AGC_GAIN_ADJ2 register is in an unknown state, causing
the AGC to behave in a way such that extra noise is added even when
level=0. The fix is to simply explicitly set MO_AGC_GAIN_ADJ2 on PCI
resume callback. We also do the same in the probe callback for
consistency.